User modified and client and freelancer moved to aggregate
