+++ src/contexts/CoreContext/domain/interfaces/repositories/IClientRepository.ts
@@ -1,5 +1,5 @@
 import { IRepository } from "@/contexts/Shared/domain/repository/IRepository";
-import { Client } from "../../entities/Client";
+import { Client } from "../../aggregates/Client";
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export interface IClientRepository extends IRepository<Client> {}
